Initializes Firebase in separate script, useEffects the FirebaseUI on Login page
