+++ frontend/src/index.js
@@ -8,9 +8,6 @@ import { ThemeProvider } from "styled-components";
 
 import theme from "./assets/theme";
 
-//Firebase logic import
-import Firebase from "./Firebase";
-
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <React.StrictMode>
@@ -26,6 +23,3 @@ root.render(
 // to log results (for example: reportWebVitals(console.log))
 // or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
 reportWebVitals();
-
-// Call firebase logic
-Firebase();
